# apps/daemon/pkg/toolbox/docs/swagger.yaml
definitions:
...
  PtyCreateRequest:
    properties:
      cols:
        type: integer
      cwd:
        type: string
      envs:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      lazyStart:
        description: Don't start PTY until first client connects
        type: boolean
      rows:
        type: integer
    type: object
  PtyCreateResponse:
    properties:
      sessionId:
        type: string
    required:
      - sessionId
    type: object
  PtyListResponse:
    properties:
      sessions:
        items:
          $ref: '#/definitions/PtySessionInfo'
        type: array
    required:
      - sessions
    type: object
  PtyResizeRequest:
    properties:
      cols:
        maximum: 1000
        minimum: 1
        type: integer
      rows:
        maximum: 1000
        minimum: 1
        type: integer
    required:
      - cols
      - rows
    type: object
  PtySessionInfo:
    properties:
      active:
        type: boolean
      cols:
        type: integer
      createdAt:
        type: string
      cwd:
        type: string
      envs:
        additionalProperties:
          type: string
        type: object
      id:
        type: string
      lazyStart:
        description: Whether this session uses lazy start
        type: boolean
      rows:
        type: integer
    required:
      - active
      - cols
      - createdAt
      - cwd
      - envs
      - id
      - lazyStart
      - rows
    type: object
...
  gin.H:
    additionalProperties: {}
    type: object
info:
  contact: {}
  description: "Daytona Toolbox API. The base URL comes from the sandbox's `toolboxProxyUrl`
    field (returned in sandbox DTO by the main Daytona API) plus the sandbox ID: `{toolboxProxyUrl}/{sandboxId}/{endpoint}`.
    Default for Daytona Cloud: `https://proxy.app.daytona.io/toolbox/{sandboxId}`."
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: Daytona Toolbox API
  version: v0.0.0-dev
paths:
...
  /process/pty:
    get:
      description: Get a list of all active pseudo-terminal sessions
      operationId: ListPtySessions
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PtyListResponse'
      summary: List all PTY sessions
      tags:
        - process
    post:
      consumes:
        - application/json
      description: Create a new pseudo-terminal session with specified configuration
      operationId: CreatePtySession
      parameters:
        - description: PTY session creation request
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/PtyCreateRequest'
      produces:
        - application/json
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/PtyCreateResponse'
      summary: Create a new PTY session
      tags:
        - process
  /process/pty/{sessionId}:
    delete:
      description: Delete a pseudo-terminal session and terminate its process
      operationId: DeletePtySession
      parameters:
        - description: PTY session ID
          in: path
          name: sessionId
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete a PTY session
      tags:
        - process
    get:
      description: Get detailed information about a specific pseudo-terminal session
      operationId: GetPtySession
      parameters:
        - description: PTY session ID
          in: path
          name: sessionId
          required: true
          type: string
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PtySessionInfo'
      summary: Get PTY session information
      tags:
        - process
  /process/pty/{sessionId}/connect:
    get:
      description: Establish a WebSocket connection to interact with a pseudo-terminal
        session
      operationId: ConnectPtySession
      parameters:
        - description: PTY session ID
          in: path
          name: sessionId
          required: true
          type: string
      responses:
        '101':
          description: Switching Protocols - WebSocket connection established
      summary: Connect to PTY session via WebSocket
      tags:
        - process
  /process/pty/{sessionId}/resize:
    post:
      consumes:
        - application/json
      description: Resize the terminal dimensions of a pseudo-terminal session
      operationId: ResizePtySession
      parameters:
        - description: PTY session ID
          in: path
          name: sessionId
          required: true
          type: string
        - description: Resize request with new dimensions
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/PtyResizeRequest'
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/PtySessionInfo'
      summary: Resize a PTY session
      tags:
        - process
...
schemes:
  - https
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and your API key.
    in: header
    name: Authorization
    type: apiKey
swagger: '2.0'